Best practices: lint fixes, module exports, shared fixtures, DRY cleanup
- Add __all__ exports to all src/ modules and populate src/__init__.py
- Make PARAM_BOUNDS public (was _PARAM_BOUNDS) and fix import in arbitrage.py
- Create tests/conftest.py with shared fixtures and synthetic data helpers
- Add pytest slow marker for integration tests
- Extract dashboard/components/helpers.py to reduce DRY violations
- Modernize pyproject.toml with full metadata, ruff config, pytest markers
- Improve CI with separate lint job, pip caching, Python matrix
- Fix all ruff lint errors (noqa:E402, zip strict=True, int(round()) cleanup)
- All 130 tests pass, ruff check clean

+++ dashboard/components/arbitrage_diag.py
@@ -61,7 +61,7 @@ def _render_butterfly(
             m=row["m"], sigma=row["sigma"],
         )
         g = durrleman_condition(k_grid, params)
-        dte = int(round(row["T"] * 365.25))
+        dte = round(row["T"] * 365.25)
         label = str(row.get("expiry", f"T={row['T']:.4f}"))
         is_free = diagnostics.butterfly_free.get(label, True)
